requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
spacy==3.7.2
python-dateutil==2.8.2
anthropic>=0.21.0
pydantic==2.5.0
jinja2==3.1.2 